+++ flutter_angle/lib/shared/classes.dart
@@ -1,3 +1,7 @@
+import 'package:flutter_angle/flutter_angle.dart';
+
+List disposeWebGL = [];
+
 class OpenGLException implements Exception {
   OpenGLException(this.message, this.error);
 
@@ -29,42 +33,90 @@ class ActiveInfo{
 
 class WebGLTexture {
   final dynamic id;
-  WebGLTexture(this.id);
+  WebGLTexture(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteTexture(this);
+  }
 }
 
 class Program {
   final dynamic id;
-  Program(this.id);
+  Program(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteProgram(this);
+  }
 }
 
 class Buffer {
   final dynamic id;
-  Buffer(this.id);
+  Buffer(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteBuffer(this);
+  }
 }
 
 class Renderbuffer {
   final dynamic id;
-  Renderbuffer(this.id);
+  Renderbuffer(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteRenderbuffer(this);
+  }
 }
 
 class Framebuffer{
   final dynamic id;
-  Framebuffer(this.id);
+  Framebuffer(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteFramebuffer(this);
+  }
 }
 
 class TransformFeedback{
   final dynamic id;
-  TransformFeedback(this.id);
+  TransformFeedback(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteTransformFeedback(this);
+  }
 }
 
 class VertexArrayObject{
   final dynamic id;
-  VertexArrayObject(this.id);
+  VertexArrayObject(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteVertexArray(this);
+  }
 }
 
 class WebGLShader{
   final dynamic id;
-  WebGLShader(this.id);
+  WebGLShader(this.id){
+    disposeWebGL.add(this);
+  }
+
+  void dispose(RenderingContext context){
+    context.deleteShader(this);
+  }
 }
 
 class UniformLocation{