+++ flutter_angle/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.8
+
+* Added google's ANGLE to android with the option to choose angle or opengl.
+
+## 0.0.7
+
+* Updated iOS device framework to fix xcode 16 archive bug.
+* Removed "import io.flutter.plugin.common.PluginRegistry.Registrar;" to fix android bug.
+
+## 0.0.6
+
+* Upgraded android for compatibility with AGP <4.2.
+
 ## 0.0.5
 
 * Added iOS simulator support.